__constants.py__ Added constants to allow 'and' and 'or' conditions, and to map these and comparisons to negated versions.
__compiler.py__
Added numerous explanatory comments and tentative TODO's.
Defined Instruction and _Label classes and used them to streamline substitution of line numbers at the final stage of compilation, which opens the door to line-number changes before that, e.g., in an optimization pass.
Added compiler.conditional_jump() as a unified location for code used in conditional jumps, using either positive or negative tests.
Added support for 'and' and 'or' in conditional jumps, both lazily evaluated as in Python.
Used the above to more efficiently compile if and while, and to allow while to use the same range of conditions that if can.

__test_all.py__
Updated if and while tests to expect more efficient compilations.
Added new tests for complex_while, while_and, and while_or.

__sample_functions.py__
This a work in progress and can be ignored for now.