compile() now accepts python functions as code argument
if f is a python function defined in a file, compile( f ) now uses inspect.getsource() to load the current source code for f from file, bypasses the def-line and docstring (if present), and compiles the body of that function.

This is demonstrated in test_all.py's new test_inspect_function_source().  It may make sense to rearticulate all other/future tests with this approach, as it is much cleaner than a textwrap.dedent'ed """-string